backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
psycopg2-binary
celery
redis
python-multipart
requests
sentence-transformers
faiss-cpu
transformers[torch]
pdfminer.six
python-docx
spacy
weasyprint
pytest
httpx
pytest-mock
